Renamed File Explorer to File Manager
With the name 'File Explorer', it seems like SkiffOS is a clone of windows or a wannabe Windows. This is not how I believe it should be.
